requirements_dev.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.3.0
joblib>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pytest>=7.4.0
mlflow>=3.8.1

# OpenTelemetry for metrics
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-prometheus>=0.43b0
prometheus-client>=0.19.0